Add files via upload
This is a small DataCamp project to apply the skills learned in the courses "Introduction to Python" and "Intermediate Python". The goal of this project is to discover if Netflix’s movies are getting shorter over time and which guest stars appear in the most popular episode of "The Office".